+++ Assets/UniGLTF/Editor/UniGLTF/gltfAssetPostprocessor.cs
@@ -44,8 +44,9 @@ namespace UniGLTF
                 return;
             }
 
-            var context = new ImporterContext();
-            context.Parse(src);
+            var parser = new GltfParser();
+            parser.ParsePath(src);
+            var context = new ImporterContext(parser);
 
             // Extract textures to assets folder
             context.ExtractImages(prefabPath);
