+++ Assets/UniGLTF/Runtime/UniGLTF/IO/GltfParser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae3083233546b58449904dca65e76005
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
