+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContextExtensions.cs
@@ -6,27 +6,28 @@ namespace UniGLTF
 {
     public static class ImporterContextExtensions
     {
-        /// <summary>
-        /// ReadAllBytes, Parse, Create GameObject
-        /// </summary>
-        /// <param name="path">allbytes</param>
-        public static void Load(this ImporterContext self, string path)
-        {
-            var bytes = File.ReadAllBytes(path);
-            self.Load(path, bytes);
-        }
+        // /// <summary>
+        // /// ReadAllBytes, Parse, Create GameObject
+        // /// </summary>
+        // /// <param name="path">allbytes</param>
+        // public static void Load(this ImporterContext self, string path)
+        // {
+        //     var bytes = File.ReadAllBytes(path);
+        //     self.Load(path, bytes);
+        // }
 
-        /// <summary>
-        /// Parse, Create GameObject
-        /// </summary>
-        /// <param name="path">gltf or glb path</param>
-        /// <param name="bytes">allbytes</param>
-        public static void Load(this ImporterContext self, string path, byte[] bytes)
-        {
-            self.Parse(path, bytes);
-            self.Load();
-            self.Root.name = Path.GetFileNameWithoutExtension(path);
-        }
+        // /// <summary>
+        // /// Parse, Create GameObject
+        // /// </summary>
+        // /// <param name="path">gltf or glb path</param>
+        // /// <param name="bytes">allbytes</param>
+        // public static void Load(this ImporterContext self, string path, byte[] bytes)
+        // {
+        //     var parser = new GltfParser();
+        //     parser.Parse(path, bytes);
+        //     self.Load(parser);
+        //     self.Root.name = Path.GetFileNameWithoutExtension(path);
+        // }
 
         /// <summary>
         /// Build unity objects from parsed gltf
