+++ Assets/VRM10.Samples/Runtime/ViewerUI.cs
@@ -333,9 +333,11 @@ namespace UniVRM10.Samples
 
                 case ".glb":
                     {
-                        var context = new UniGLTF.ImporterContext();
                         var file = File.ReadAllBytes(path);
-                        context.ParseGlb(file);
+                        var parser = new GltfParser();
+                        parser.ParseGlb(file);
+
+                        var context = new UniGLTF.ImporterContext(parser);
                         context.Load();
                         context.ShowMeshes();
                         context.EnableUpdateWhenOffscreen();
@@ -347,8 +349,10 @@ namespace UniVRM10.Samples
                 case ".gltf":
                 case ".zip":
                     {
-                        var context = new UniGLTF.ImporterContext();
-                        context.Parse(path);
+                        var parser = new GltfParser();
+                        parser.ParsePath(path);
+
+                        var context = new UniGLTF.ImporterContext(parser);
                         context.Load();
                         context.ShowMeshes();
                         context.EnableUpdateWhenOffscreen();
